fix(efcore): Skip/Take pagination and Uint64 SQL literals (#661)

+++ src/EFCore.Ydb/src/Query/Internal/YdbQuerySqlGenerator.cs
@@ -80,7 +80,7 @@ public class YdbQuerySqlGenerator(QuerySqlGeneratorDependencies dependencies) : 
         else
         {
             // We must specify number here because offset without limit leads to exception
-            Sql.Append(ulong.MaxValue.ToString());
+            Sql.Append(int.MaxValue.ToString());
         }
 
         // ReSharper disable once InvertIf
